AirLibrary::Security

Struct SecureStorage

pub struct SecureStorage {
    credentials: Arc<RwLock<HashMap<String, EncryptedCredential>>>,
    master_key: SecureBytes,
    key_version: u32,
    auditor: SecurityAuditor,
}

Secure credential storage with AES-GCM encryption

Fields
credentials: Arc<RwLock<HashMap<String, EncryptedCredential>>>

Encrypted credentials storage

master_key: SecureBytes

Master key for encryption/decryption (zeroized on drop)

key_version: u32

Key version for key rotation support

auditor: SecurityAuditor

Security auditor

Implementations

impl SecureStorage

pub fn New(master_key: Vec<u8>, auditor: SecurityAuditor) -> Self

Create a new secure storage with a master key

pub fn DeriveKeyFromPassword( password: &str, salt: Option<&[u8]>, ) -> (Vec<u8>, [u8; 16])

Generate a secure master key from password using PBKDF2

pub async fn Store(&self, key: &str, credential: &str) -> Result<()>

Store a credential encrypted with AES-GCM

pub async fn Retrieve(&self, key: &str) -> Result<Option<String>>

Retrieve and decrypt a credential

fn EncryptCredential( &self, data: &str, nonce: &[u8; 12], salt: &[u8; 16], ) -> Result<String>

Encrypt credential data using AES-GCM

fn DecryptCredential( &self, cipher_text: &str, nonce: &[u8], salt: &[u8], ) -> Result<String>

Decrypt credential data

fn DeriveSubkey(&self, salt: &[u8]) -> Result<SecureBytes>

Derive a subkey from the master key using PBKDF2

pub async fn RotateMasterKey( &self, new_master_key: Vec<u8>, ) -> Result<KeyRotationResult>

Rotate the master key and re-encrypt all credentials

pub async fn ClearAll(&self) -> Result<()>

Clear all stored credentials

pub async fn CredentialCount(&self) -> usize

Get the number of stored credentials

pub async fn ListCredentials(&self) -> Vec<String>

List all credential keys (without exposing credentials)