AirLibrary::Tracing

Struct TraceGenerator

pub struct TraceGenerator {
    trace_spans: Arc<RwLock<HashMap<String, TraceSpan>>>,
    sampling_config: Arc<RwLock<SamplingConfig>>,
}

Trace ID generator and manager with sampling support

Fields
trace_spans: Arc<RwLock<HashMap<String, TraceSpan>>>sampling_config: Arc<RwLock<SamplingConfig>>
Implementations

impl TraceGenerator

pub fn new() -> Self

Create a new trace generator with default sampling

pub fn with_sampling(sampling_config: SamplingConfig) -> Result<Self>

Create a new trace generator with custom sampling

pub fn generate_trace_id() -> String

Generate a new trace ID with panic recovery

pub fn generate_span_id() -> String

Generate a new span ID

pub async fn should_sample(&self, is_critical: bool) -> bool

Determine if a trace should be sampled based on configuration

pub fn parse_trace_context(header: &str) -> Result<PropagationContext>

Parse W3C Trace Context header

pub async fn create_span( &self, trace_id: String, operation_name: impl Into<String>, parent_span_id: Option<String>, attributes: Option<HashMap<String, String>>, ) -> Result<TraceSpan>

Create a new trace span with optional sampling check

pub async fn add_span_event( &self, span_id: &str, event_name: impl Into<String>, attributes: HashMap<String, String>, ) -> Result<()>

Add an event to a span

pub async fn mark_span_active(&self, span_id: &str) -> Result<()>

Mark a span as active

pub async fn complete_span( &self, span_id: &str, error: Option<String>, ) -> Result<u64>

Complete a span with optional error

pub async fn add_span_attribute( &self, span_id: &str, key: String, value: String, ) -> Result<()>

Add attribute to a span

pub async fn add_span_attributes( &self, span_id: &str, attributes: HashMap<String, String>, ) -> Result<()>

Add multiple attributes to a span

pub async fn get_span(&self, span_id: &str) -> Result<TraceSpan>

Get a span by ID

pub async fn get_trace_spans(&self, trace_id: &str) -> Result<Vec<TraceSpan>>

Get all spans for a trace

pub async fn get_trace_metadata(&self, trace_id: &str) -> Result<TraceMetadata>

Get trace metadata

pub async fn export_trace(&self, trace_id: &str) -> Result<String>

Export trace in JSON format

pub async fn cleanup_old_spans( &self, older_than_ms: Option<u64>, ) -> Result<usize>

Clean up old spans (older than specified milliseconds)

pub async fn get_statistics(&self) -> TraceStatistics

Get trace statistics

fn sanitize_attributes( &self, attributes: HashMap<String, String>, ) -> HashMap<String, String>

Sanitize attributes to remove sensitive data

fn sanitize_error_message(&self, message: &str) -> String

Sanitize error messages to remove sensitive data